WIP: Move countingNamingFunction to Naming
